Refactor User Profile Page: Modularized components into ProfileSidebar, ProfileContent, and DevXNavbar. Introduced ContractCard for contract display and improved state management for user data and sidebar navigation.
